requirements.txt
streamlit>=1.38.0
pydantic>=2.0.0
PyMuPDF>=1.24.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
requests>=2.32.0
duckduckgo_search>=6.3.0
numpy>=1.26.0
scikit-learn>=1.5.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
